name_from_urlの場所がおかしかったとおもうのでうつした
